Refactor package structure to include 'api' subpackage.
Moved classes and interfaces to the 'api' subpackage for better organization and clarity. This change affects event, link, model, preset, and view packages and ensures a more modular structure.
