+++ plugin/src/main/java/net/thenextlvl/worlds/WorldsPlugin.java
@@ -10,13 +10,13 @@ import net.kyori.adventure.text.minimessage.tag.resolver.Placeholder;
 import net.kyori.adventure.text.minimessage.tag.resolver.TagResolver;
 import net.thenextlvl.worlds.command.WorldCommand;
 import net.thenextlvl.worlds.controller.WorldLinkController;
-import net.thenextlvl.worlds.view.GeneratorView;
-import net.thenextlvl.worlds.link.LinkController;
+import net.thenextlvl.worlds.api.view.GeneratorView;
+import net.thenextlvl.worlds.api.link.LinkController;
 import net.thenextlvl.worlds.listener.PortalListener;
 import net.thenextlvl.worlds.listener.ServerListener;
-import net.thenextlvl.worlds.preset.Presets;
+import net.thenextlvl.worlds.api.preset.Presets;
 import net.thenextlvl.worlds.version.PluginVersionChecker;
-import net.thenextlvl.worlds.view.LevelView;
+import net.thenextlvl.worlds.api.view.LevelView;
 import net.thenextlvl.worlds.view.PaperLevelView;
 import net.thenextlvl.worlds.view.PluginGeneratorView;
 import org.bstats.bukkit.Metrics;
