+++ plugin/src/main/java/net/thenextlvl/worlds/view/PaperLevelView.java
@@ -5,13 +5,14 @@ import core.io.PathIO;
 import core.nbt.file.NBTFile;
 import core.nbt.tag.*;
 import lombok.RequiredArgsConstructor;
+import net.thenextlvl.worlds.api.view.LevelView;
 import net.thenextlvl.worlds.WorldsPlugin;
-import net.thenextlvl.worlds.model.LevelExtras;
-import net.thenextlvl.worlds.model.WorldPreset;
-import net.thenextlvl.worlds.preset.Biome;
-import net.thenextlvl.worlds.preset.Layer;
-import net.thenextlvl.worlds.preset.Preset;
-import net.thenextlvl.worlds.preset.Structure;
+import net.thenextlvl.worlds.api.model.LevelExtras;
+import net.thenextlvl.worlds.api.model.WorldPreset;
+import net.thenextlvl.worlds.api.preset.Biome;
+import net.thenextlvl.worlds.api.preset.Layer;
+import net.thenextlvl.worlds.api.preset.Preset;
+import net.thenextlvl.worlds.api.preset.Structure;
 import org.bukkit.NamespacedKey;
 import org.bukkit.World;
 import org.bukkit.WorldCreator;
